Split out lag tests too
